+++ crates/biome_cli/src/commands/mod.rs
@@ -50,6 +50,9 @@ pub enum BiomeCommand {
         /// Prints the Biome daemon server logs
         #[bpaf(long("daemon-logs"), switch)]
         bool,
+        /// Prints the Biome configuration that the applied linter rules
+        #[bpaf(long("linter"), switch)]
+        bool,
     ),
     /// Start the Biome daemon server process
     #[bpaf(command)]