+++ crates/biome_cli/src/commands/rage.rs
@@ -12,9 +12,14 @@ use tokio::runtime::Runtime;
 use crate::commands::daemon::read_most_recent_log_file;
 use crate::service::enumerate_pipes;
 use crate::{service, CliDiagnostic, CliSession, VERSION};
+use serde::Serialize;
 
 /// Handler for the `rage` command
-pub(crate) fn rage(session: CliSession, daemon_logs: bool) -> Result<(), CliDiagnostic> {
+pub(crate) fn rage(
+    session: CliSession,
+    daemon_logs: bool,
+    linter: bool,
+) -> Result<(), CliDiagnostic> {
     let terminal_supports_colors = termcolor::BufferWriter::stdout(ColorChoice::Auto)
         .buffer()
         .supports_color();
@@ -35,7 +40,7 @@ pub(crate) fn rage(session: CliSession, daemon_logs: bool) -> Result<(), CliDiag
     {EnvVarOs("JS_RUNTIME_NAME")}
     {EnvVarOs("NODE_PACKAGE_MANAGER")}
 
-    {RageConfiguration(&session.app.fs)}
+    {RageConfiguration(&session.app.fs, linter)}
     {WorkspaceRage(session.app.workspace.deref())}
     ));
 
@@ -165,7 +170,7 @@ impl Display for RunningRomeServer {
     }
 }
 
-struct RageConfiguration<'a, 'app>(&'a DynRef<'app, dyn FileSystem>);
+struct RageConfiguration<'a, 'app>(&'a DynRef<'app, dyn FileSystem>, bool);
 
 impl Display for RageConfiguration<'_, '_> {
     fn fmt(&self, fmt: &mut Formatter) -> io::Result<()> {
@@ -195,13 +200,35 @@ impl Display for RageConfiguration<'_, '_> {
                         markup!(<Dim>"Loaded successfully"</Dim>)
                     };
 
-                    markup! (
+                    // Print linter rules if --linter option is true
+                    if self.1 {
+                        markup! (
+                            {KeyValuePair("Status", status)}
+                            {KeyValuePair("Formatter disabled", markup!({DebugDisplay(configuration.is_formatter_disabled())}))}
+                            {KeyValuePair("Linter disabled", markup!({DebugDisplay(configuration.is_linter_disabled())}))}
+                            {KeyValuePair("Organize imports disabled", markup!({DebugDisplay(configuration.is_organize_imports_disabled())}))}
+                            {KeyValuePair("VCS disabled", markup!({DebugDisplay(configuration.is_vcs_disabled())}))}
+                            {Section("Linter rules")}
+                            {KeyValuePair("Recommend", markup!({DebugDisplay(configuration.get_linter_rules().recommended.unwrap_or(false))}))}
+                            {KeyValuePair("All", markup!({DebugDisplay(configuration.get_linter_rules().all.unwrap_or(false))}))}
+                            {KeyValuePair("A11y", markup!({DebugDisplayJson(configuration.get_linter_rules().a11y)}))}
+                            {KeyValuePair("Complexity", markup!({DebugDisplayJson(configuration.get_linter_rules().complexity)}))}
+                            {KeyValuePair("Correctness", markup!({DebugDisplayJson(configuration.get_linter_rules().correctness)}))}
+                            {KeyValuePair("Nursery", markup!({DebugDisplayJson(configuration.get_linter_rules().nursery)}))}
+                            {KeyValuePair("Performance", markup!({DebugDisplayJson(configuration.get_linter_rules().performance)}))}
+                            {KeyValuePair("Security", markup!({DebugDisplayJson(configuration.get_linter_rules().security)}))}
+                            {KeyValuePair("Style", markup!({DebugDisplayJson(configuration.get_linter_rules().style)}))}
+                            {KeyValuePair("Suspicious", markup!({DebugDisplayJson(configuration.get_linter_rules().suspicious)}))}
+                        ).fmt(fmt)?
+                    } else {
+                        markup! (
                     {KeyValuePair("Status", status)}
                     {KeyValuePair("Formatter disabled", markup!({DebugDisplay(configuration.is_formatter_disabled())}))}
                     {KeyValuePair("Linter disabled", markup!({DebugDisplay(configuration.is_linter_disabled())}))}
                     {KeyValuePair("Organize imports disabled", markup!({DebugDisplay(configuration.is_organize_imports_disabled())}))}
                     {KeyValuePair("VCS disabled", markup!({DebugDisplay(configuration.is_vcs_disabled())}))}
                 ).fmt(fmt)?
+                    }
                 }
             }
             Err(err) => markup! (
@@ -223,6 +250,17 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> io::Result<()> {
         write!(f, "{:?}", self.0)
+    }
+}
+
+struct DebugDisplayJson<T>(T);
+
+impl<T> Display for DebugDisplayJson<T>
+where
+    T: Serialize,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> io::Result<()> {
+        write!(f, "{}", serde_json::to_string(&self.0).unwrap())
     }
 }